+++ tests/test_data_structures.py
@@ -87,7 +87,6 @@ class TestCompressionParameters(unittest.TestCase):
         self.assertEqual(p.ldm_bucket_size_log, 7)
 
         p = zstd.ZstdCompressionParameters(ldm_hash_rate_log=8)
-        self.assertEqual(p.ldm_hash_every_log, 8)
         self.assertEqual(p.ldm_hash_rate_log, 8)
 
     def test_estimated_compression_context_size(self):
@@ -119,18 +118,8 @@ class TestCompressionParameters(unittest.TestCase):
         self.assertEqual(p.compression_strategy, 3)
 
     def test_ldm_hash_rate_log(self):
-        with self.assertRaisesRegex(
-            ValueError, "cannot specify both ldm_hash_rate_log"
-        ):
-            zstd.ZstdCompressionParameters(
-                ldm_hash_rate_log=8, ldm_hash_every_log=4
-            )
-
         p = zstd.ZstdCompressionParameters(ldm_hash_rate_log=8)
-        self.assertEqual(p.ldm_hash_every_log, 8)
-
-        p = zstd.ZstdCompressionParameters(ldm_hash_every_log=16)
-        self.assertEqual(p.ldm_hash_every_log, 16)
+        self.assertEqual(p.ldm_hash_rate_log, 8)
 
     def test_overlap_log(self):
         p = zstd.ZstdCompressionParameters(overlap_log=2)